+++ CHANGELOG.md
@@ -4,8 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Soul gem drop rates receive a small bonus from looting
 ### Fixed
 - Soul gems not having their data on the client when playing on a server [#380]
+- Soul gem drops crashing the game in some cases [#380]
 - Hardened stone generating over the ocean [#379]
 
 ## [3.2.1] - 2019-07-03