+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.2.2] - 2019-07-05
 ### Changed
 - Soul gem drop rates receive a small bonus from looting
+- Fluffy puffs and fluffy puff seeds can be placed in the composter
 ### Fixed
 - Soul gems not having their data on the client when playing on a server [#380]
 - Soul gem drops crashing the game in some cases [#380]